Merge in MB/mb-backend from bugfix-MB-392 to develop

* commit '54d6aab9ec83bd8ec66a840c5b6b72e9ee520b23':
  bugfix-MB-392 пофиксил проценты и добавил Page

+++ app/src/main/java/pro/mbroker/app/service/impl/CalculatorServiceImpl.java
@@ -203,8 +203,8 @@ public class CalculatorServiceImpl implements CalculatorService {
                 mortgageSum.compareTo(creditProgram.getCreditParameter().getMaxMortgageSum()) <= 0 &&
                 creditProgram.getCreditParameter().getMinCreditTerm() <= creditTermMonths &&
                 creditProgram.getCreditParameter().getMaxCreditTerm() >= creditTermMonths &&
-                downPaymentPercentage >= (creditProgram.getCreditParameter().getMinDownPayment()) &&
-                downPaymentPercentage <= (creditProgram.getCreditParameter().getMaxDownPayment()) &&
+                downPaymentPercentage >= creditProgram.getCreditParameter().getMinDownPayment().intValue() &&
+                downPaymentPercentage <= creditProgram.getCreditParameter().getMaxDownPayment().intValue() &&
                 isRegionEligible(request, creditProgram) &&
                 isMaternalCapital(request, creditProgram);
     }
